requirements.txt
fastapi
uvicorn[standard]
asyncpg
asyncio
ariadne
ariadne-graphql-modules
python-dotenv
numpy
pandas
valkey
psycopg2-binary
